requirements.txt
setuptools~=57.0.0
numpy~=1.22.0
gym~=0.21.0
Pillow~=9.0.0
tensorflow~=2.7.0
matplotlib~=3.5.1